Fix: 결과페이지 컴포넌트 분리 #52
