+++ src/app/content/[id]/result/_ContentResultPage.scss
@@ -10,7 +10,7 @@
   & > section {
     scroll-margin: 6rem;
     scroll-snap-align: start;
-    height: 90svh;
+    min-height: 90svh;
   }
 
   .heading-2 {
@@ -31,144 +31,5 @@
     font-size: 1.8rem;
     line-height: auto;
     color: $color-gray-3;
-  }
-}
-
-.theme {
-  @include flexColumn;
-  background-color: $color-yellow-1;
-  .bubble-list {
-    @include flexColumn;
-    flex: 1;
-    gap: 0.5rem;
-    justify-content: center;
-  }
-  .bubble-item {
-    @include flexColumn;
-    gap: 0.5rem;
-    padding: 3rem 5.2rem;
-    width: 80%;
-    max-width: 72rem;
-    min-width: 30rem;
-    position: relative;
-
-    backdrop-filter: blur(10px);
-    background-color: rgba($color-white, 0.6);
-    box-shadow: 0px 7px 15px 5px rgba($color-yellow-2, 0.6);
-    color: $color-gray-5;
-
-    &:nth-of-type(even) {
-      margin-left: auto;
-      border-radius: 70px 70px 70px 0px;
-    }
-    &:nth-of-type(odd) {
-      margin-right: auto;
-      border-radius: 70px 70px 0px 70px;
-    }
-
-    &:nth-of-type(2) {
-      margin-top: -2.4rem;
-    }
-    &:nth-of-type(3) {
-      width: 90%;
-      margin-top: -2rem;
-      margin-left: auto;
-      margin-right: auto;
-
-      padding-left: 12rem;
-    }
-  }
-  .bubble-title {
-    margin-bottom: 0.8rem;
-    font-size: 2rem;
-    font-weight: 600;
-  }
-  .bubble-text {
-    font-size: 1.6rem;
-    line-height: 2.4rem;
-  }
-}
-.top-rated-talker {
-  @include flexColumn;
-  justify-content: center;
-
-  background-color: #fffaed;
-
-  &::before {
-    content: '';
-    margin-bottom: -7rem;
-    height: 16rem;
-    width: 100%;
-    display: block;
-    background-image: url('/image/wave.svg');
-    background-color: $color-yellow-1;
-    background-size: 16rem;
-    background-repeat: repeat-x;
-    background-position: top left;
-  }
-
-  .graph {
-    flex: 1;
-    @include flexCenter;
-    flex-wrap: wrap;
-    gap: 5px;
-  }
-
-  .rank-list {
-    max-width: 40rem;
-    flex: 1;
-    .rank-item {
-      @include flexCenter;
-      // display: flex;
-      // align-items: center;
-      padding: 0.6rem 2rem;
-      border-radius: 1.4rem;
-      color: $color-gray-5;
-      font-weight: 600;
-      font-size: 2rem;
-
-      &:first-of-type {
-        transform: scale(1);
-        background-color: $color-yellow-2;
-      }
-      &:not(:first-of-type) {
-        transform: scale(0.9);
-        margin-top: 0.3rem;
-      }
-
-      & > span {
-        min-width: fit-content;
-      }
-
-      .rank {
-        margin-right: 1rem;
-      }
-      .name {
-        flex: 1;
-      }
-      .count {
-        margin-right: 1.6rem;
-        font-size: 1.6rem;
-        font-weight: 500;
-        opacity: 0.7;
-      }
-      .percentage {
-      }
-    }
-  }
-}
-.mbti {
-  .mbti-list {
-  }
-  .mbti-item {
-  }
-}
-.most-popular-word {
-  @include flexColumn;
-
-  .graph {
-    flex: 1;
-    height: auto;
-    @include flexCenter;
   }
 }
