+++ src/app/content/[id]/result/components/MbtiSection/_MbtiSection.scss
@@ -1,0 +1,9 @@
+@import '../../../../../../styles/variable';
+@import '../../../../../../styles/mixins/flex.scss';
+
+.mbti {
+  .mbti-list {
+  }
+  .mbti-item {
+  }
+}
